test: share e2e navigation helpers
