+++ frontend/e2e/text-editing.test.ts
@@ -1,18 +1,5 @@
-import { expect, openEditor, test } from "./fixtures";
+import { expect, openEditor, scrollToStep, test } from "./fixtures";
 import type { Page, Locator } from "@playwright/test";
-
-function photos(page: Page): Locator {
-  return page.locator('[role="button"][aria-pressed]');
-}
-
-async function scrollToStep(page: Page, country: string, stepName: string) {
-  const nav = page.getByRole("navigation");
-  await nav.getByText(country).click();
-  const step = nav.getByText(stepName);
-  await expect(step).toBeVisible({ timeout: 3_000 });
-  await step.click();
-  await expect(photos(page).first()).toBeVisible({ timeout: 5_000 });
-}
 
 /** Find the first step name textbox (aria-label="Step name") in the viewer. */
 function stepNameTextbox(page: Page): Locator {
