+++ frontend/e2e/undo-redo.test.ts
@@ -1,9 +1,15 @@
-import { expect, openEditor, test } from "./fixtures";
+import {
+  expect,
+  openEditor,
+  photoButtons,
+  scrollToStep,
+  test,
+} from "./fixtures";
 import type { Page, Locator } from "@playwright/test";
 import { PHOTO_SHORTCUTS } from "../src/composables/shortcutKeys";
 
 function photos(page: Page): Locator {
-  return page.locator('[role="button"][aria-pressed]');
+  return photoButtons(page);
 }
 
 /** The unused photo count badge in the inspector panel. */
@@ -12,15 +18,6 @@ function unusedBadge(page: Page): Locator {
     .getByRole("complementary")
     .filter({ hasText: "Unused" })
     .getByText(/^\d+$/);
-}
-
-async function scrollToStep(page: Page, country: string, stepName: string) {
-  const nav = page.getByRole("navigation");
-  await nav.getByText(country).click();
-  const step = nav.getByText(stepName);
-  await expect(step).toBeVisible({ timeout: 3_000 });
-  await step.click();
-  await expect(photos(page).first()).toBeVisible({ timeout: 5_000 });
 }
 
 test.describe("Undo & redo", () => {
